Clean random_forest sources and add tutorial
